requirements.txt
anthropic==0.49.0
python-dotenv==1.0.1
python-docx==1.1.0
gradio==5.23.3
pandas==2.2.1
openpyxl==3.1.2
PyPDF2==3.0.1
xlrd>=2.0.1